requirements.txt
django==3.1.3
djangorestframework==3.12.2              
python-decouple==3.4
django-solo==1.1.3
cryptography==3.3
flake8
celery==4.3.1
psycopg2==2.8.6
# sslchecker
pyopenssl
json2html
requests==2.25.1
django-admin-autocomplete-filter==0.6.1
django-admin-rangefilter==0.6.4